Move the creation and download log file logic to separate module
